requirements.txt
fastmcp
markdownify
readabilipy
httpx
pdfminer.six
python-docx
pydantic

